refactor(cli): derive the standalone help section from requires_wallet
The wallet-free command set was stated three times: the requires_wallet
match, the hand-listed standalone_commands array, and the pinning test
that held the two in agreement. The array also needed its placeholder
fields rebuilt whenever a standalone command gained an argument.

The set now has the single statement requires_wallet. The tests'
one-sample-per-variant list moves into the module as every_command,
held complete by the existing set-equality pin, and a
wallet_free_commands helper filters it, so format_help's section split
derives from the predicate. The standalone_commands array and its
agreement test dissolve. The change was driven test-first: a new pin,
red at the build until the helper existed, asserts the rendered
standalone section equals the derived wallet-free names. The findings
record marks the classification residue resolved.

Both feature configurations pass cargo nextest (197 and 204 tests),
and clippy and rustfmt are clean.
